acp-agents/agents/travel_concierge_agent/pyproject.toml
[project]
name = "travel-concierge-agent"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "acp-sdk>=0.9.1",
    "crewai[tools]==0.120.1",
    "fastapi>=0.115.12",
    "pydantic-ai>=0.2.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = ["ruff>=0.11.10"]
